Add support for excluding assemblies and honor alternative found versions
asdf
