+++ AsmSpy.CommandLine/Utilities/XmlWriterExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Xml;
 
-namespace AsmSpy.CommandLine
+namespace AsmSpy.CommandLine.Utilities
 {
     public static class XmlWriterExtensions
     {
