+++ AsmSpy.Core/VisualizerOptions.cs
@@ -1,18 +1,12 @@
-﻿using System;
+﻿using System.Collections.Generic;
 
 namespace AsmSpy.Core
 {
     public class VisualizerOptions
     {
-        public bool SkipSystem { get; }
-        public bool OnlyConflicts { get; }
-        public string ReferencedStartsWith { get; }
-
-        public VisualizerOptions(bool skipSystem, bool onlyConflicts, string referencedStartsWith)
-        {
-            SkipSystem = skipSystem;
-            OnlyConflicts = onlyConflicts;
-            ReferencedStartsWith = referencedStartsWith ?? throw new ArgumentNullException(nameof(referencedStartsWith));
-        }
+        public bool SkipSystem { get; set; }
+        public bool OnlyConflicts { get; set; }
+        public string ReferencedStartsWith { get; set; }
+        public IList<string> Exclude { get; set; }
     }
 }
