+++ README.md
@@ -33,6 +33,7 @@ It will output a list of all conflicting assembly references. That is where diff
 | dot | Export dependency graph to a [DOT](https://en.wikipedia.org/wiki/DOT_(graph_description_language)) file.<br> Supported formats:  -dt \<filename\> |
 | xml | Export dependency graph to a xml file.<br> Supported formats:  -x \<filename\> |
 | rsw | Will only analyze assemblies if their referenced assemblies starts with the given value.<br> Supported formats:  -rsw \<string\>, --referencedstartswith \<string\> |
+| e | Will exclude assemblies if they start with the given value. This option can be provided multiple times.<br> Supported formats:  -e \<string\>, --exclude \<string\> |
 | tree | Write a dependency tree to the console.<br>Supported formats: -tr --tree |
 | i | Include subdirectories in search.<br> Supported formats:  -i, --includesub |
 | c | Use the binding redirects of the given configuration file (Web.config or App.config) <br> Supported formats: -c \<string>, --configurationFile \<string> |
